encontros/13_airflow/requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-amazon==8.10.0
